feat(loader): better filename defaults by env
Match webpack standards and don't include a content hash unless the filename specifies it. Add a
useCache control to the icon-sprite-loader. End result is that hashing and caching is opt-in for
production envs.

BREAKING CHANGE:
The sprite filename is no longer hashed by default and must be opted in to